Use DebugInfo from ModelTrackerInterface to retrieve synchronized data.
Previous approach was to use a separate grabber to retrieve
images from the server. However, this approach does not allow
to get synchronized cMo + image. Change to use the debugging data
structure provided in the ModelTrackerInterface.

This approach has a drawback: the grabbing client is not used
anymore to grab in ActionTrackingMbt and ActionDisplayMbt.