Commit skeleton for tracking with command law computation.
The idea of this class is to instrument a tracking (whatever its type)
with a command law computation.
A command law only needs to compute a velocity and a direction
from a cMo and a desired cMo so it should not be related to a particular
algorithm.

The design pattern decorator seems particularly suited to implement
this behavior.

Status: to be completed.